NP SSML Classification Code
TOF-Times is a two-stage semi-supervised machine learning analysis scheme used for the classification of nanoparticle data from spICP-TOFMS. This analysis scheme utilizes a model-guided classification scheme by creating two semi-supervised machine learning models. The first machine learning model is used to identify false-positive predictions as a result of variability in multi-elemental particle signatures from spICP-TOFMS as a function of critical masses. The second model is retrained with the addition of unclassifiable particle classes for the false-positive predictions. This two-stage approach creates a more robust classification model and reduces the impact of mass distribution bias and variable multi-elemental signatures for NP classification.